Working on the PowerShell module cloning part. Fixes #14.
Moved the Windows Update command to the top, which appears to be required
to properly install VS (in some cases). And it gets a bunch of reboots out
of the way early.

Maybe we'll have git on the command line by the time we run the powershell
module commands. We'll see!