notepad2-mod: fix uninstallation
The code that calls Get-Uninstaller expects it to return exactly one
value, the path to the uninstaller executable.

1) As Get-Uninstaller searches multiple registry locations, in theory it
could return multiple paths. Fix this so that only the first one is
returned.

2) Applications (including notepad2-mod) often follow good practices and
write the UninstallString value with the path to the uninstaller
executable quoted. Such path cannot be used directly - the quotes must be
skipped.

Implement rudimentary parsing logic to extract the path to the executable
if it is quoted, otherwise return the full UninstallString, as before.
This logic does not handle some cases that would work correctly with
Programs and Features (e.g. unquoted path followed by arguments), but
Notepad2-mod does not hit those cases, so the logic is sufficient here.