Fix for race condition bug in Chocolaty 0.10.3 on community edition.
In 0.10.3, the Chocolaty auto uninstaller is attempting to uninstall the package even though
it has already been uninstalled.  This occurs when it the auto-uninstaller fails to find the
recently removed installer.  This change adds a short wait to chocolateyUninstall.ps1 to
allow time for the changes to propagate in the registry.

This bug is fixed in Chocolaty 0.10.4, where it will no longer fail when the uninstaller is
not found

see:
https://github.com/chocolatey/choco/issues/1035